Integration tests added. All of them pass successfully
