+++ src/main/java/com/cursor/exceptions/IncorrectIdException.java
@@ -1,6 +1,6 @@
 package com.cursor.exceptions;
 
-public class IncorrectIdException extends Throwable {
+public class IncorrectIdException extends Exception {
     public IncorrectIdException(String message) {
         super(message);
     }
