+++ src/main/java/com/cursor/exceptions/IncorrectMovieDtoException.java
@@ -1,6 +1,6 @@
 package com.cursor.exceptions;
 
-public class IncorrectMovieDtoException extends Throwable {
+public class IncorrectMovieDtoException extends Exception {
     public IncorrectMovieDtoException(String message) {
         super(message);
     }
