+++ src/main/java/com/cursor/exceptions/IncorrectRateException.java
@@ -1,6 +1,6 @@
 package com.cursor.exceptions;
 
-public class IncorrectRateException extends Throwable {
+public class IncorrectRateException extends Exception {
     public IncorrectRateException(String message) {
         super(message);
     }
