+++ src/main/resources/application.properties
@@ -7,10 +7,8 @@ spring.datasource.password=12345678
 #Hibernate
 
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL95Dialect
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create
+# i've added cascade for the entity Rate. That's why you need to use "create" here to delete previous schema and create new
+# after that you can change its value to "update" if you want
 
 spring.main.allow-bean-definition-overriding=true
-
-
-
-
